Reorganized test to split common, macOS and iOS
